Organize files into directories: client, common, server, test
