+++ packages/ddp-client/client/stream_client_sockjs.js
@@ -1,12 +1,12 @@
-import { DDP, LivedataTest } from "./namespace.js";
 import { _ } from 'meteor/underscore';
 import { Meteor } from 'meteor/meteor';
-import { toSockjsUrl } from './urlHelpers';
 
 // This populates a global variable
 import './sockjs-0.3.4';
 
-import { addCommonMethodsToPrototype } from './stream_client_common';
+import { DDP, LivedataTest } from "../common/namespace.js";
+import { toSockjsUrl } from '../common/urlHelpers';
+import { addCommonMethodsToPrototype } from '../common/stream_client_common';
 
 // @param url {String} URL to Meteor app
 //   "http://subdomain.meteor.com/" or "/" or
