+++ packages/ddp-client/common/namespace.js
