--- packages/ddp-client/server.js
@@ -1,7 +1,0 @@
-import { DDP, LivedataTest } from './namespace';
-
-import './stream_client_nodejs';
-
-import './livedata_connection';
-
-export { DDP, LivedataTest };
