+++ packages/ddp-client/server/server.js
@@ -1,0 +1,7 @@
+import { DDP, LivedataTest } from '../common/namespace';
+
+import './stream_client_nodejs';
+
+import '../common/livedata_connection';
+
+export { DDP, LivedataTest };
