+++ packages/ddp-client/test/stream_tests.js
@@ -1,4 +1,4 @@
-import { LivedataTest } from "./namespace.js";
+import { LivedataTest } from "../common/namespace.js";
 
 Tinytest.add("stream - status", function (test) {
   // Very basic test. Just see that it runs and returns something. Not a
