+++ packages/ddp-client/test/stub_stream.js
